test(core): split output hook tests
